+++ src/noise_model/noise_channel.cpp
@@ -5,6 +5,7 @@
 #include "qristal/core/tools/zip_tool.hpp"
 
 #include <numeric>
+#include <random>
 #include <unsupported/Eigen/KroneckerProduct>
 #include "qristal/core/primitives.hpp"
 
@@ -431,5 +432,287 @@ namespace qristal
 
       return kraus_ops;
     }
+
+    //============================================ Process matrix interpolation methods ============================================
+
+    Eigen::MatrixXcd createIdealU3ProcessMatrix(double theta, double phi, double lambda) {
+      // Calculate I, X, Y, and Z components of U3 gate
+      std::complex<double> U_I = 0.5 * cos(theta/2) * (1.0 + exp(std::complex(0.0, phi + lambda)));
+      std::complex<double> U_X = 0.5 * sin(theta/2) * (exp(std::complex(0.0, phi)) - exp(std::complex(0.0, lambda)));
+      std::complex<double> U_Y = std::complex(0.0, -0.5) * sin(theta/2) * (exp(std::complex(0.0, phi)) + exp(std::complex(0.0, lambda)));
+      std::complex<double> U_Z = 0.5 * cos(theta/2) * (1.0 - exp(std::complex(0.0, phi + lambda)));
+
+      // Build 1-qubit process matrix from components.
+      Eigen::MatrixXcd U(4, 4);
+      U << U_I * std::conj(U_I), U_I * std::conj(U_X), U_I * std::conj(U_Y), U_I * std::conj(U_Z),
+           U_X * std::conj(U_I), U_X * std::conj(U_X), U_X * std::conj(U_Y), U_X * std::conj(U_Z),
+           U_Y * std::conj(U_I), U_Y * std::conj(U_X), U_Y * std::conj(U_Y), U_Y * std::conj(U_Z),
+           U_Z * std::conj(U_I), U_Z * std::conj(U_X), U_Z * std::conj(U_Y), U_Z * std::conj(U_Z);
+
+      return U;
+    }
+
+    Eigen::MatrixXcd create1QubitNoisyProcessMatrix(double theta, double phi, double lambda,
+        Eigen::VectorXd channel_params) {
+      // Create ideal process matrix
+      Eigen::MatrixXcd ideal_process_mat = qristal::createIdealU3ProcessMatrix(theta, phi, lambda);
+      // Convert process matrix to superoperator representation
+      Eigen::MatrixXcd ideal_process_mat_super = qristal::process_to_superoperator(ideal_process_mat);
+
+      // The Kraus matrices in the noise channels are independent of the qubit index, so we can directly create
+      // one channel, say for qubit 0, and apply them to all qubits.
+      double gen_amp_damp_rate = channel_params(0);
+      double gen_phase_damp_rate = channel_params(1);
+      double depol_1qubit_rate = channel_params(2);
+      qristal::NoiseChannel gen_phase_amp_damp =
+          qristal::GeneralizedPhaseAmplitudeDampingChannel::Create(0, 0.0, gen_amp_damp_rate, gen_phase_damp_rate);
+      qristal::NoiseChannel depol_1qubit = qristal::DepolarizingChannel::Create(0, depol_1qubit_rate);
+
+      std::vector<Eigen::MatrixXcd> channel_kraus_matrices;
+      for (const auto &x : gen_phase_amp_damp) {
+        channel_kraus_matrices.emplace_back(matrix_to_eigen(x.matrix));
+      }
+      for (const auto &x : depol_1qubit) {
+        channel_kraus_matrices.emplace_back(matrix_to_eigen(x.matrix));
+      }
+
+      // Convert noise channel matrices to superoperator representation
+      Eigen::MatrixXcd channel_kraus_matrices_super = qristal::kraus_to_superoperator(channel_kraus_matrices);
+      // Apply noise channel on ideal process matrix
+      Eigen::MatrixXcd noisy_process_mat_super = channel_kraus_matrices_super * ideal_process_mat_super;
+
+      return noisy_process_mat_super;
+    }
+
+    Eigen::MatrixXcd createNQubitNoisyProcessMatrix(size_t nb_qubits,
+        std::vector<double> theta, std::vector<double> phi, std::vector<double> lambda,
+        Eigen::VectorXd channel_params) {
+      // Generate noisy process matrix by creating an ideal process matrix and acting the noise channel
+      // matrices on it.
+      Eigen::MatrixXcd process_matrix = Eigen::MatrixXcd::Zero(std::pow(2, 2*nb_qubits), std::pow(2, 2*nb_qubits));
+      for (size_t i = 0; i < nb_qubits; i++) {
+        // Create noisy 1-qubit process matrix
+        double gen_amp_damp_rate = channel_params(i);
+        double gen_phase_damp_rate = channel_params(i + nb_qubits);
+        double depol_1qubit_rate = channel_params(i + 2*nb_qubits);
+        Eigen::VectorXd channel_params_1qubit(3);
+        channel_params_1qubit << gen_amp_damp_rate, gen_phase_damp_rate, depol_1qubit_rate;
+        Eigen::MatrixXcd noisy_process_mat_super =  qristal::create1QubitNoisyProcessMatrix(
+            theta[i], phi[i], lambda[i], channel_params_1qubit);
+
+        // Expand the vector space of the 1-qubit process matrix N-qubit vector space
+        Eigen::MatrixXcd process_mat_tmp;
+        for (size_t qId = 0; qId < nb_qubits; qId++) {
+          Eigen::MatrixXcd Id_left = Eigen::MatrixXcd::Identity(std::pow(2, 2*qId), std::pow(2, 2*qId));
+          Eigen::MatrixXcd Id_right = Eigen::MatrixXcd::Identity(std::pow(2, 2*(nb_qubits - qId - 1)), std::pow(2, 2*(nb_qubits - qId - 1)));
+          process_mat_tmp = Eigen::kroneckerProduct(Id_left, noisy_process_mat_super).eval();
+          process_mat_tmp = Eigen::kroneckerProduct(process_mat_tmp, Id_right).eval();
+        }
+        process_matrix += process_mat_tmp;
+      }
+
+      return process_matrix;
+    }
+
+    Eigen::VectorXd processMatrixSolverInternal(size_t nb_qubits, size_t nb_params, size_t max_iter,
+        Eigen::LevenbergMarquardt<Eigen::NumericalDiff<qristal::LMFunctorNoisy>, double> lm,
+        std::optional<Eigen::VectorXd> guess_params) {
+      static std::random_device rd;
+      static std::mt19937 gen(rd());
+      // Choose some physically meaningful random values as guess values
+      static std::uniform_real_distribution<double> dist_amp_damp(0.0, 2e-6);
+      static std::uniform_real_distribution<double> dist_phase_damp(0.0, 2e-3);
+      static std::uniform_real_distribution<double> dist_depol1(0.0, 2e-5);
+      // Random values to perturb the solved parameters
+      static std::uniform_real_distribution<double> dist_perturb(-0.1, 0.1);
+
+      // Solve for guess parameters x's.
+      // Do one some rough iterations first to get a converged solution and to get some "previous"
+      // output to be fed into subsequent iterations
+      lm.fnorm = std::numeric_limits<double>::signaling_NaN();
+      double sum_fvec = std::numeric_limits<double>::signaling_NaN();
+      int info = 10;
+      Eigen::VectorXd x(nb_params);
+      bool x_is_positive_valued = 0;
+      for (size_t i = 0; i < max_iter; i++) {
+        if (std::isnan(lm.fnorm) || std::isnan(sum_fvec) || info > 5 || !x_is_positive_valued) {
+          if (guess_params) { // Use guess parameters if provided
+            x = guess_params.value();
+          } else { // Use random guess parameters
+            std::vector<double> gen_amp_damp_rate_vec, gen_phase_damp_rate_vec, depol_1qubit_rate_vec;
+            for (size_t i = 0; i < nb_qubits; i++) {
+              gen_amp_damp_rate_vec.emplace_back(dist_amp_damp(gen));
+              gen_phase_damp_rate_vec.emplace_back(dist_phase_damp(gen));
+              depol_1qubit_rate_vec.emplace_back(dist_depol1(gen));
+            }
+            Eigen::VectorXd gen_amp_damp_rate = Eigen::Map<Eigen::VectorXd, Eigen::Unaligned>(gen_amp_damp_rate_vec.data(), gen_amp_damp_rate_vec.size());
+            Eigen::VectorXd gen_phase_damp_rate = Eigen::Map<Eigen::VectorXd, Eigen::Unaligned>(gen_phase_damp_rate_vec.data(), gen_phase_damp_rate_vec.size());
+            Eigen::VectorXd depol_1qubit_rate = Eigen::Map<Eigen::VectorXd, Eigen::Unaligned>(depol_1qubit_rate_vec.data(), depol_1qubit_rate_vec.size());
+            x << gen_amp_damp_rate, gen_phase_damp_rate, depol_1qubit_rate;
+          }
+
+          info = lm.minimize(x);
+          sum_fvec = std::accumulate(lm.fvec.begin(), lm.fvec.end(), 0.0);
+          for (size_t i = 0; i < x.size(); i++) {
+            if (x[i] < 0) {
+              x_is_positive_valued = 0;
+              break;
+            } else {
+              x_is_positive_valued = 1;
+            }
+          }
+        } else {
+          break;
+        }
+      }
+      Eigen::VectorXd x_prev = x;
+      double sum_fvec_prev = sum_fvec;
+      double fnorm_prev = lm.fnorm;
+
+      // Now do the subsequent iterations for convergence
+      for (size_t i = 0; i < max_iter; i++) {
+        int info = lm.minimize(x);
+        double sum_fvec = std::accumulate(lm.fvec.begin(), lm.fvec.end(), 0.0);
+        bool x_is_positive_valued = 1;
+        for (size_t i = 0; i < x.size(); i++) {
+          if (x[i] < 0) {
+            x_is_positive_valued = 0;
+            break;
+          }
+        }
+
+        // Perturb small angles if the norm of current vector function (fnorm) is not converged to < 1e-4
+        // The tolerance value of 1e-4 is selected from the reconstruction of the process matrix (reconstructed_mat)
+        // gives an acceptable difference relative to the initial process matrix (process_mat_noisy).
+        if (lm.fnorm > 1e-4 || sum_fvec > 1e-6 || info > 5 || !x_is_positive_valued) {
+          if (sum_fvec > sum_fvec_prev) {
+            x = x_prev;
+          } else {
+            sum_fvec_prev = sum_fvec;
+            x_prev = x;
+          }
+
+          for (size_t i = 0; i < x.size(); i++) {
+            x[i] = (x[i] < 0) ? x[i] * -1.0 : x[i];
+            assert(x[i] >= 0);
+            x[i] += dist_perturb(gen) * x[i];
+          }
+        } else {
+          break;
+        }
+      }
+
+      return x;
+    }
+
+    Eigen::VectorXd processMatrixSolver1Qubit(Eigen::MatrixXcd process_matrix,
+        double theta, double phi, double lambda, size_t max_iter,
+        size_t maxfev, double xtol, double ftol, double gtol) {
+      size_t nb_params = 3; // 3 parameters for the 2 noise channels
+
+      // To use Eigen's numerical differentiation, we need to convert the complex matrix into a complex vector.
+      Eigen::VectorXcd process_vec = Eigen::Map<Eigen::VectorXcd>(process_matrix.data(), process_matrix.rows() * process_matrix.cols());
+
+      qristal::LMFunctorNoisy functor;
+      functor.input_vec = process_vec;
+      functor.m = process_vec.size();
+      functor.n = nb_params;
+      functor.nb_qubits = 1;
+      functor.theta = {theta};
+      functor.phi = {phi};
+      functor.lambda = {lambda};
+
+      Eigen::NumericalDiff<qristal::LMFunctorNoisy> numDiff(functor);
+      Eigen::LevenbergMarquardt<Eigen::NumericalDiff<qristal::LMFunctorNoisy>, double> lm(numDiff);
+      lm.parameters.maxfev = maxfev;
+      lm.parameters.xtol = xtol;
+      lm.parameters.ftol = ftol;
+      lm.parameters.gtol = gtol;
+      // Solve for guess parameters x's.
+      Eigen::VectorXd x = qristal::processMatrixSolverInternal(1, nb_params, max_iter, lm);
+
+      return x;
+    }
+
+    Eigen::VectorXd processMatrixSolverNQubit(std::vector<Eigen::MatrixXcd> process_matrix_1qubit,
+        Eigen::MatrixXcd process_matrix_Nqubit, size_t nb_qubits,
+        std::vector<double> theta, std::vector<double> phi, std::vector<double> lambda, size_t max_iter,
+        size_t maxfev, double xtol, double ftol, double gtol) {
+      //--------------------------------- Solve the 1-qubit process matrices ---------------------------------
+      // Solve the 1-qubit process matrices and use their solutions as a guess to the N-qubit process matrix
+      assert(nb_qubits == process_matrix_1qubit.size());
+      std::vector<double> gen_amp_damp_rate_vec, gen_phase_damp_rate_vec, depol_1qubit_rate_vec; // Vectors so store 1-qubit solutions
+      for (size_t i = 0; i < nb_qubits; i++) {
+        Eigen::VectorXd x = qristal::processMatrixSolver1Qubit(process_matrix_1qubit[i], theta[i], phi[i], lambda[i], max_iter);
+        // Collect solutions to be used as guess values for the N-qubit process matrix solver
+        gen_amp_damp_rate_vec.emplace_back(x(0));
+        gen_phase_damp_rate_vec.emplace_back(x(1));
+        depol_1qubit_rate_vec.emplace_back(x(2));
+      }
+
+      //--------------------------------- Solve the N-qubit process matrix ---------------------------------
+      size_t nb_params = 3 * nb_qubits; // 3 parameters for the 2 noise channels, per qubit
+      Eigen::VectorXd gen_amp_damp_rate = Eigen::Map<Eigen::VectorXd, Eigen::Unaligned>(gen_amp_damp_rate_vec.data(), gen_amp_damp_rate_vec.size());
+      Eigen::VectorXd gen_phase_damp_rate = Eigen::Map<Eigen::VectorXd, Eigen::Unaligned>(gen_phase_damp_rate_vec.data(), gen_phase_damp_rate_vec.size());
+      Eigen::VectorXd depol_1qubit_rate = Eigen::Map<Eigen::VectorXd, Eigen::Unaligned>(depol_1qubit_rate_vec.data(), depol_1qubit_rate_vec.size());
+      Eigen::VectorXd x_guess(nb_params);
+      x_guess << gen_amp_damp_rate, gen_phase_damp_rate, depol_1qubit_rate;
+
+      // To use Eigen's numerical differentiation, we need to convert the complex matrix into a complex vector.
+      Eigen::VectorXcd process_vec = Eigen::Map<Eigen::VectorXcd>(process_matrix_Nqubit.data(), process_matrix_Nqubit.rows() * process_matrix_Nqubit.cols());
+
+      qristal::LMFunctorNoisy functor;
+      functor.input_vec = process_vec;
+      functor.m = process_vec.size();
+      functor.n = nb_params;
+      functor.nb_qubits = nb_qubits;
+      functor.theta = theta;
+      functor.phi = phi;
+      functor.lambda = lambda;
+
+      Eigen::NumericalDiff<qristal::LMFunctorNoisy> numDiff(functor);
+      Eigen::LevenbergMarquardt<Eigen::NumericalDiff<qristal::LMFunctorNoisy>, double> lm(numDiff);
+      lm.parameters.maxfev = maxfev;
+      lm.parameters.xtol = xtol;
+      lm.parameters.ftol = ftol;
+      lm.parameters.gtol = gtol;
+      // Solve for guess parameters x's.
+      Eigen::VectorXd x = qristal::processMatrixSolverInternal(nb_qubits, nb_params, max_iter, lm, x_guess);
+
+      return x;
+    }
+
+    Eigen::MatrixXcd processMatrixInterpolator(size_t nb_qubits,
+        std::vector<Eigen::MatrixXcd> process_matrix_1qubit_1, std::vector<Eigen::MatrixXcd> process_matrix_1qubit_2,
+        Eigen::MatrixXcd process_matrix_Nqubit_1, Eigen::MatrixXcd process_matrix_Nqubit_2,
+        std::vector<double> theta1, std::vector<double> phi1, std::vector<double> lambda1,
+        std::vector<double> theta2, std::vector<double> phi2, std::vector<double> lambda2,
+        double theta_target, double phi_target, double lambda_target, size_t max_iter, size_t maxfev,
+        double xtol, double ftol, double gtol) {
+
+      // Solve for the noise channel parameters at angle {tx1, ty1, tz1}
+      Eigen::VectorXd params1 = qristal::processMatrixSolverNQubit(
+          process_matrix_1qubit_1, process_matrix_Nqubit_1, nb_qubits, theta1, phi1, lambda1,
+          max_iter, maxfev, xtol, ftol, gtol);
+      // Solve for the noise channel parameters at angle {tx2, ty2, tz2}
+      Eigen::VectorXd params2 = qristal::processMatrixSolverNQubit(
+          process_matrix_1qubit_2, process_matrix_Nqubit_2, nb_qubits, theta2, phi2, lambda2,
+          max_iter, maxfev, xtol, ftol, gtol);
+
+      // Calculate average of each param in param1 and param2
+      Eigen::VectorXd params_avg = (params1 + params2) / 2;
+
+      // Create output process matrix using average noise channel parameters of both
+      // input process matrices
+      std::vector<double> theta_target_vec, phi_target_vec, lambda_target_vec;
+      for (size_t i = 0; i < nb_qubits; i++) {
+        theta_target_vec.emplace_back(theta_target);
+        phi_target_vec.emplace_back(phi_target);
+        lambda_target_vec.emplace_back(lambda_target);
+      }
+      Eigen::MatrixXcd output_process_matrix = qristal::createNQubitNoisyProcessMatrix(
+          nb_qubits, theta_target_vec, phi_target_vec, lambda_target_vec, params_avg);
+
+      return output_process_matrix;
+    }
 }
-